#80760 add LayerPeriod Trex endpoint. Implement dummy designName in list.

+++ src/service/TRex/src/netstandard/VSS.TRex.SiteModels/SiteModelMachineDesign.cs
@@ -9,7 +9,7 @@ namespace VSS.TRex.SiteModels
   /// </summary>
   public class SiteModelMachineDesign : ISiteModelMachineDesign
   {
-    private const int VERSION_NUMBER = 1;
+    private const byte VERSION_NUMBER = 1;
 
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
